Develop

+++ src/main/java/com/nawell/magasin/repositories/ClientRepository.java
@@ -3,5 +3,7 @@ package com.nawell.magasin.repositories;
 import com.nawell.magasin.models.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+import java.util.Optional;
+
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
