+++ src/main/java/com/nawell/magasin/repositories/CommandeRepository.java
@@ -1,7 +1,10 @@
 package com.nawell.magasin.repositories;
 
+import com.nawell.magasin.models.Client;
 import com.nawell.magasin.models.Commande;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommandeRepository extends JpaRepository<Commande, Integer> {
+import java.util.Optional;
+
+public interface CommandeRepository extends JpaRepository<Commande, Long> {
 }
